+++ package.json
@@ -21,8 +21,8 @@
     }
   },
   "scripts": {
-    "preprebuild": "bolt ws exec -- rimraf dist",
-    "prebuild": "ts-node tools/link-ts.ts && bolt ws exec -- node_modules/.bin/tsc --emitDeclarationOnly",
+    "clean": "bolt ws exec -- rimraf dist",
+    "prebuild": "bolt clean && ts-node tools/link-ts.ts && bolt ws exec -- node_modules/.bin/tsc --emitDeclarationOnly",
     "build": "bolt ws exec -- node_modules/.bin/babel src -d dist/src --quiet --extensions \".ts\"",
     "postbuild": "ts-node tools/test-dist",
     "commit": "git-cz",
@@ -83,10 +83,14 @@
     "@types/cross-spawn": "^6.0.0",
     "@types/debug": "^0.0.30",
     "@types/electron-winstaller": "^2.6.1",
+    "@types/form-data": "^2.2.1",
     "@types/fs-extra": "^5.0.2",
+    "@types/lodash.merge": "^4.6.3",
     "@types/log-symbols": "^2.0.0",
+    "@types/mime-types": "^2.1.0",
     "@types/mocha": "^5.1.0",
     "@types/node": "^9.6.5",
+    "@types/node-fetch": "^1.6.8",
     "@types/ora": "^1.3.4",
     "@types/pify": "^3.0.1",
     "@types/proxyquire": "^1.3.28",
