+++ packages/maker/appx/package.json
@@ -19,6 +19,7 @@
   },
   "dependencies": {
     "@electron-forge/maker-base": "6.0.0-beta.3",
+    "@electron-forge/shared-types": "^6.0.0-beta.3",
     "cross-spawn": "^6.0.4",
     "fs-extra": "^5.0.0",
     "parse-author": "^2.0.0",
