+++ packages/maker/deb/package.json
@@ -21,7 +21,8 @@
     "node": ">= 6.0"
   },
   "dependencies": {
-    "@electron-forge/maker-base": "6.0.0-beta.3"
+    "@electron-forge/maker-base": "6.0.0-beta.3",
+    "@electron-forge/shared-types": "^6.0.0-beta.3"
   },
   "optionalDependencies": {
     "electron-installer-debian": "^0.8.0"
