+++ packages/maker/dmg/package.json
@@ -22,6 +22,7 @@
   },
   "dependencies": {
     "@electron-forge/maker-base": "6.0.0-beta.3",
+    "@electron-forge/shared-types": "^6.0.0-beta.3",
     "fs-extra": "^5.0.0",
     "pify": "^3.0.0"
   },
