+++ packages/maker/rpm/package.json
@@ -22,6 +22,7 @@
   },
   "dependencies": {
     "@electron-forge/maker-base": "6.0.0-beta.3",
+    "@electron-forge/shared-types": "^6.0.0-beta.3",
     "pify": "^3.0.0"
   },
   "optionalDependencies": {
