+++ packages/maker/wix/package.json
@@ -19,6 +19,7 @@
   },
   "dependencies": {
     "@electron-forge/maker-base": "6.0.0-beta.3",
+    "@electron-forge/shared-types": "^6.0.0-beta.3",
     "electron-wix-msi": "^1.3.0"
   }
 }
