+++ packages/publisher/electron-release-server/src/Config.ts
@@ -1,0 +1,6 @@
+export interface PublisherERSConfig {
+  baseUrl: string;
+  username: string;
+  password: string;
+  channel?: string;
+}
