+++ packages/publisher/github/src/Config.ts
@@ -1,0 +1,12 @@
+import { Options } from '@octokit/rest';
+
+export interface PublisherGithubConfig {
+  repository: {
+    name: string;
+    owner: string;
+    draft?: boolean;
+  }
+  authToken?: string;
+  octokitOptions?: Options;
+  prerelease?: boolean;
+}
