+++ packages/publisher/s3/package.json
@@ -20,6 +20,7 @@
   "dependencies": {
     "@electron-forge/async-ora": "^6.0.0-beta.3",
     "@electron-forge/publisher-base": "6.0.0-beta.3",
+    "@electron-forge/shared-types": "^6.0.0-beta.3",
     "aws-sdk": "^2.9.0",
     "debug": "^3.0.0",
     "s3": "^4.4.0"
